Added depth limits to breadth first and depth first search
a parameter can be applied to an applyBreadthFirstSearch or applyDepthFirstSearch method that specifies the maximum depth of a node that will be checked to see if it is the goal state.
This could be useful for applying a search algorithm on something where we know the maximum possible depth of the solution. (e.g. a rubik's cube where the maximum possible depth is 20)

Also added a custom exception that is thrown if there is no solution found up to the maximum depth.
Also refactored the code