Removed node pruning from depth first search due to an inherent error
Pruning with depth first search with a depth limit (which is always present in theory, in practise that limit is Integer.MAX_INT) makes it possible for a solution to be impossible to find even if there is one. This is because if a node is expanded it is always added to the list of nodes to check whether to prune its children, however this means if a solution is not found beneath this node and a node higher up in the search tree produces a matching node it will not expand it.

Possible Fixes:
Recursion (eww) (but doable)
Wait until all nodes at a certain depth have been expanded before adding those nodes to the list of expanded nodes. (big space complexity horrors here[I think it s O(x^n)])